pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/minions"]

[project]
name = "open-minions"
version = "0.1.0"
description = "One-shot, end-to-end coding agents - unattended agents that turn tasks into pull requests"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Open Minions" }]
keywords = ["ai", "coding", "agent", "llm", "mcp", "automation", "github", "slack"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
]
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "mcp>=1.0.0",
    "rich>=13.9.0",
    "typer>=0.12.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
slack = [
    "slack-bolt>=1.20.0",
    "slack-sdk>=3.30.0",
]
github = [
    # GitHub API is handled via httpx (already a core dep)
]
all = [
    "open-minions[slack]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "respx>=0.21.0",
]

[project.scripts]
minion = "minions.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
